adding unit tests

+++ app/src/main/java/com/example/moviesdb/domain/model/MediaItem.kt
@@ -4,7 +4,7 @@ data class MediaItem(
     val id: Int?,
     val title: String?,
     val releaseYear: String?,
-    val rating: Float?,
+    val rating: Double?,
     val posterImageUrl: String?,
     val mediaType: MediaType,
 )
